Clean up makefile and occamygen
